backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==45.0.6
dataclasses-json==0.6.7
distro==1.9.0
docx2pdf==0.1.8
easygui==0.98.3
ebcdic==1.1.1
et_xmlfile==2.0.0
extract-msg==0.54.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fpdf==1.7.2
frozenlist==1.7.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.13
lark==1.1.9
lxml==6.0.0
marshmallow==3.26.1
msoffcrypto-tool==5.4.2
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
olefile==0.47
oletools==0.60.2
openai==1.99.3
openpyxl==3.1.5
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pcodedmp==1.2.6
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
red-black-tree-mod==1.22
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
RTFDE==0.1.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
win_unicode_console==0.5
yarl==1.20.1
zstandard==0.23.0
